requirements.txt
papermill==2.3.4
pydantic==1.9.0
typer==0.4.0
pyyaml==6.0
